Moved all partitioner logic into partitioner.py
